refactor: split zero-job-detail.ts into domain modules
Split the 937-line zero-job-detail.ts into 8 domain-specific modules
under signals/zero-page/job-detail/. The original file becomes a barrel
re-export so all existing import paths continue to work unchanged.

Modules: agent-name, detail, instructions, settings, connectors,
firewall, schedule, delete, plus an index barrel with the temporary
fetchZeroJobData$ orchestrator.

Closes #7776
