+++ turbo/apps/platform/src/signals/zero-page/zero-job-detail.ts
@@ -1,937 +1,1 @@
-import { command, computed, state } from "ccstate";
-import { toast } from "@vm0/ui/components/ui/sonner";
-import {
-  zeroAgentsByIdContract,
-  zeroAgentInstructionsContract,
-  zeroSchedulesMainContract,
-  zeroSchedulesByNameContract,
-  zeroSchedulesEnableContract,
-  zeroUserConnectorsContract,
-  type FirewallPolicies,
-} from "@vm0/core";
-import { throwIfAbort } from "../utils.ts";
-import { logger } from "../log.ts";
-import { zeroClient$ } from "../api-client.ts";
-import { search } from "../location.ts";
-import type { AgentDetail, AgentInstructions } from "./agent-types.ts";
-import {
-  buildCronExpression,
-  buildAtTime,
-  isAtTimePast,
-  type ScheduleBody,
-  type CronTimeOption,
-} from "./cron.ts";
-import { reloadAgents$ } from "./agents-list.ts";
-import type { ScheduleEntry } from "../../views/zero-page/zero-schedule-card.tsx";
-
-const L = logger("ZeroJobDetail");
-
-// ---------------------------------------------------------------------------
-// Agent name — set when navigating to a subagent detail page
-// ---------------------------------------------------------------------------
-
-const internalAgentName$ = state<string | null>(null);
-const setZeroJobAgentName$ = command(({ set }, name: string | null) => {
-  set(internalAgentName$, name);
-});
-
-// ---------------------------------------------------------------------------
-// Active tab
-// ---------------------------------------------------------------------------
-
-function isValidTab(tab: string): boolean {
-  return (
-    tab === "authorization" ||
-    tab === "schedule" ||
-    tab === "profile" ||
-    tab === "instructions"
-  );
-}
-
-function getInitialTab(): string {
-  const params = new URLSearchParams(search());
-  const tab = params.get("tab") ?? "";
-  return isValidTab(tab) ? tab : "authorization";
-}
-
-const internalActiveTab$ = state("authorization");
-
-export const zeroJobActiveTab$ = computed((get) => {
-  return get(internalActiveTab$);
-});
-
-export const setZeroJobActiveTab$ = command(({ set }, tab: string) => {
-  set(internalActiveTab$, tab);
-  const url = new URL(location.href);
-  if (tab === "authorization") {
-    url.searchParams.delete("tab");
-  } else {
-    url.searchParams.set("tab", tab);
-  }
-  history.replaceState(null, "", url.toString());
-});
-
-// ---------------------------------------------------------------------------
-// Agent detail
-// ---------------------------------------------------------------------------
-
-interface ZeroJobDetailState {
-  detail: AgentDetail | null;
-  loading: boolean;
-  error: string | null;
-}
-
-const detailState$ = state<ZeroJobDetailState>({
-  detail: null,
-  loading: false,
-  error: null,
-});
-
-export const zeroJobDetail$ = computed((get) => {
-  return get(detailState$).detail;
-});
-export const zeroJobDetailLoading$ = computed((get) => {
-  return get(detailState$).loading;
-});
-export const zeroJobDetailError$ = computed((get) => {
-  return get(detailState$).error;
-});
-
-const fetchZeroJobDetail$ = command(
-  async ({ get, set }, _signal: AbortSignal) => {
-    const name = get(internalAgentName$);
-    if (!name) {
-      return;
-    }
-
-    set(detailState$, (prev) => {
-      return { ...prev, loading: true, error: null };
-    });
-
-    try {
-      const client = get(zeroClient$)(zeroAgentsByIdContract);
-      const result = await client.get({ params: { id: name } });
-      if (result.status !== 200) {
-        throw new Error(`Failed to fetch agent (${result.status})`);
-      }
-
-      set(detailState$, {
-        detail: result.body,
-        loading: false,
-        error: null,
-      });
-    } catch (error) {
-      throwIfAbort(error);
-      L.error("Failed to fetch agent detail:", error);
-      set(detailState$, (prev) => {
-        return {
-          ...prev,
-          loading: false,
-          error: error instanceof Error ? error.message : "Unknown error",
-        };
-      });
-    }
-  },
-);
-
-// ---------------------------------------------------------------------------
-// Agent instructions
-// ---------------------------------------------------------------------------
-
-interface ZeroJobInstructionsState {
-  instructions: AgentInstructions | null;
-  loading: boolean;
-  error: string | null;
-}
-
-const instructionsState$ = state<ZeroJobInstructionsState>({
-  instructions: null,
-  loading: false,
-  error: null,
-});
-
-export const zeroJobInstructions$ = computed((get) => {
-  return get(instructionsState$).instructions;
-});
-export const zeroJobInstructionsLoading$ = computed((get) => {
-  return get(instructionsState$).loading;
-});
-export const zeroJobInstructionsError$ = computed((get) => {
-  return get(instructionsState$).error;
-});
-
-const fetchZeroJobInstructions$ = command(
-  async ({ get, set }, _signal: AbortSignal) => {
-    const detail = get(zeroJobDetail$);
-    if (!detail) {
-      return;
-    }
-
-    set(instructionsState$, { instructions: null, loading: true, error: null });
-
-    try {
-      const client = get(zeroClient$)(zeroAgentInstructionsContract);
-      const result = await client.get({ params: { id: detail.agentId } });
-      if (result.status !== 200) {
-        throw new Error(`Failed to fetch instructions (${result.status})`);
-      }
-
-      set(instructionsState$, {
-        instructions: result.body,
-        loading: false,
-        error: null,
-      });
-    } catch (error) {
-      throwIfAbort(error);
-      L.error("Failed to fetch instructions:", error);
-      set(instructionsState$, {
-        instructions: null,
-        loading: false,
-        error:
-          error instanceof Error
-            ? error.message
-            : "Failed to load instructions",
-      });
-    }
-  },
-);
-
-// ---------------------------------------------------------------------------
-// Instructions editing
-// ---------------------------------------------------------------------------
-
-const editedContent$ = state<string | null>(null);
-
-export const zeroJobEditedContent$ = computed((get) => {
-  return get(editedContent$);
-});
-
-export const zeroJobInstructionsDirty$ = computed((get) => {
-  const edited = get(editedContent$);
-  const instructions = get(instructionsState$).instructions;
-  const savedBody = instructions?.content ?? "";
-  return edited !== null && edited.trim() !== savedBody.trim();
-});
-
-export const setZeroJobEditedContent$ = command(({ set }, value: string) => {
-  set(editedContent$, value);
-});
-
-export const discardZeroJobEdit$ = command(({ set }) => {
-  set(editedContent$, null);
-});
-
-const jobBuilding$ = state(false);
-export const zeroJobBuilding$ = computed((get) => {
-  return get(jobBuilding$);
-});
-
-const internalBuildError$ = state<string | null>(null);
-export const zeroJobBuildError$ = computed((get) => {
-  return get(internalBuildError$);
-});
-
-export const buildZeroJobInstructions$ = command(
-  async ({ get, set }, signal: AbortSignal) => {
-    const detail = get(zeroJobDetail$);
-    const raw = get(editedContent$);
-    if (!detail?.agentId || raw === null) {
-      return;
-    }
-    const edited = raw.trim();
-
-    set(jobBuilding$, true);
-    set(internalBuildError$, null);
-
-    try {
-      const client = get(zeroClient$)(zeroAgentInstructionsContract);
-      const result = await client.update({
-        params: { id: detail.agentId },
-        body: { content: edited },
-      });
-      signal.throwIfAborted();
-
-      if (result.status !== 200) {
-        const errorDetail =
-          result.status === 401 ||
-          result.status === 403 ||
-          result.status === 404 ||
-          result.status === 422
-            ? result.body.error.message
-            : `status ${result.status}`;
-        throw new Error(`Build failed: ${errorDetail}`);
-      }
-
-      // Clear building state before content updates so the editor remounts
-      // with editable=true. Setting jobBuilding$, instructionsState$, and
-      // editedContent$ in the same synchronous block batches them into a
-      // single render — the editor key changes and the new editor starts
-      // editable immediately.
-      set(jobBuilding$, false);
-
-      const current = get(instructionsState$).instructions;
-      set(instructionsState$, {
-        instructions: { content: edited, filename: current?.filename ?? null },
-        loading: false,
-        error: null,
-      });
-
-      set(editedContent$, null);
-      await set(fetchZeroJobDetail$, signal);
-      toast.success("Instructions saved");
-    } catch (error) {
-      throwIfAbort(error);
-      set(
-        internalBuildError$,
-        "Failed to build instructions. Please try again.",
-      );
-    } finally {
-      set(jobBuilding$, false);
-    }
-  },
-);
-
-// ---------------------------------------------------------------------------
-// Settings: update agent metadata (displayName, sound)
-// ---------------------------------------------------------------------------
-
-const internalSaving$ = state(false);
-export const zeroJobSettingsSaving$ = computed((get) => {
-  return get(internalSaving$);
-});
-
-interface ZeroJobSettingsUpdate {
-  displayName?: string;
-  description?: string;
-  sound?: string;
-  avatarUrl?: string | null;
-}
-
-export const zeroJobUpdateSettings$ = command(
-  async ({ get, set }, update: ZeroJobSettingsUpdate, signal: AbortSignal) => {
-    const detail = get(zeroJobDetail$);
-    if (!detail) {
-      throw new Error("No compose detail found");
-    }
-
-    set(internalSaving$, true);
-    try {
-      const client = get(zeroClient$)(zeroAgentsByIdContract);
-      const result = await client.updateMetadata({
-        params: { id: detail.agentId },
-        body: update,
-      });
-      signal.throwIfAborted();
-      if (result.status !== 200) {
-        const detail =
-          result.status === 401 ||
-          result.status === 403 ||
-          result.status === 404
-            ? result.body.error.message
-            : `status ${result.status}`;
-        throw new Error(`Save failed: ${detail}`);
-      }
-
-      await set(fetchZeroJobDetail$, signal);
-      set(reloadAgents$);
-      toast.success("Profile saved");
-    } catch (error) {
-      throwIfAbort(error);
-      L.error("Failed to update settings:", error);
-      toast.error(error instanceof Error ? error.message : "Save failed");
-    } finally {
-      set(internalSaving$, false);
-    }
-  },
-);
-
-// ---------------------------------------------------------------------------
-// Connectors management — user-level permissions per agent
-// ---------------------------------------------------------------------------
-
-interface UserConnectorPermissionsState {
-  enabledTypes: string[];
-  loading: boolean;
-  error: string | null;
-}
-
-const userConnectorPermissionsState$ = state<UserConnectorPermissionsState>({
-  enabledTypes: [],
-  loading: false,
-  error: null,
-});
-
-const fetchZeroJobUserConnectors$ = command(
-  async ({ get, set }, _signal: AbortSignal) => {
-    const detail = get(zeroJobDetail$);
-    if (!detail?.agentId) {
-      return;
-    }
-
-    set(userConnectorPermissionsState$, {
-      enabledTypes: [],
-      loading: true,
-      error: null,
-    });
-
-    try {
-      const client = get(zeroClient$)(zeroUserConnectorsContract);
-      const result = await client.get({ params: { id: detail.agentId } });
-      if (result.status !== 200) {
-        throw new Error(
-          `Failed to fetch connector permissions (${result.status})`,
-        );
-      }
-      set(userConnectorPermissionsState$, {
-        enabledTypes: result.body.enabledTypes,
-        loading: false,
-        error: null,
-      });
-    } catch (error) {
-      throwIfAbort(error);
-      L.error("Failed to fetch user connector permissions:", error);
-      set(userConnectorPermissionsState$, {
-        enabledTypes: [],
-        loading: false,
-        error:
-          error instanceof Error
-            ? error.message
-            : "Failed to load connector permissions",
-      });
-    }
-  },
-);
-
-const internalAddedConnectors$ = state<string[] | null>(null);
-
-const seededConnectors$ = computed((get) => {
-  return get(userConnectorPermissionsState$).enabledTypes;
-});
-
-export const zeroJobConnectorsLoading$ = computed((get) => {
-  return get(userConnectorPermissionsState$).loading;
-});
-
-export const zeroJobAddedConnectors$ = computed((get) => {
-  const local = get(internalAddedConnectors$);
-  if (local !== null) {
-    return local;
-  }
-  return get(seededConnectors$);
-});
-
-export const zeroJobConnectorsDirty$ = computed((get) => {
-  const local = get(internalAddedConnectors$);
-  if (local === null) {
-    return false;
-  }
-  const seeded = get(seededConnectors$);
-  if (local.length !== seeded.length) {
-    return true;
-  }
-  const sorted = [...local].sort();
-  const seededSorted = [...seeded].sort();
-  return sorted.some((s, i) => {
-    return s !== seededSorted[i];
-  });
-});
-
-export const addZeroJobConnector$ = command(({ get, set }, name: string) => {
-  if (get(internalAddedConnectors$) === null) {
-    set(internalAddedConnectors$, get(seededConnectors$));
-  }
-  set(internalAddedConnectors$, (prev) => {
-    return [...(prev ?? []), name];
-  });
-});
-
-export const removeZeroJobConnector$ = command(({ get, set }, name: string) => {
-  if (get(internalAddedConnectors$) === null) {
-    set(internalAddedConnectors$, get(seededConnectors$));
-  }
-  set(internalAddedConnectors$, (prev) => {
-    return (prev ?? []).filter((s) => {
-      return s !== name;
-    });
-  });
-});
-
-export const discardZeroJobConnectors$ = command(({ set }) => {
-  set(internalAddedConnectors$, null);
-});
-
-export const saveZeroJobConnectors$ = command(
-  async ({ get, set }, signal: AbortSignal) => {
-    const detail = get(zeroJobDetail$);
-    if (!detail?.agentId) {
-      throw new Error("No agent detail loaded");
-    }
-
-    set(internalSaving$, true);
-    try {
-      const enabledTypes = get(internalAddedConnectors$) ?? [];
-      const client = get(zeroClient$)(zeroUserConnectorsContract);
-      const result = await client.update({
-        params: { id: detail.agentId },
-        body: { enabledTypes },
-      });
-      signal.throwIfAborted();
-
-      if (result.status !== 200) {
-        const errorDetail =
-          result.status === 401 ||
-          result.status === 403 ||
-          result.status === 404
-            ? result.body.error.message
-            : `status ${result.status}`;
-        throw new Error(`Save failed: ${errorDetail}`);
-      }
-
-      set(internalAddedConnectors$, null);
-      set(userConnectorPermissionsState$, {
-        enabledTypes: result.body.enabledTypes,
-        loading: false,
-        error: null,
-      });
-      toast.success("Connectors saved");
-    } catch (error) {
-      throwIfAbort(error);
-      L.error("Failed to save connectors:", error);
-      toast.error(
-        error instanceof Error ? error.message : "Failed to save connectors",
-      );
-    } finally {
-      set(internalSaving$, false);
-    }
-  },
-);
-
-// ---------------------------------------------------------------------------
-// Firewall policies — fetched from zero agents endpoint
-// ---------------------------------------------------------------------------
-
-const internalFirewallPolicies$ = state<FirewallPolicies | null>(null);
-
-export const zeroJobFirewallPolicies$ = computed((get) => {
-  return get(internalFirewallPolicies$);
-});
-
-export const setZeroJobFirewallPolicies$ = command(
-  ({ set }, policies: FirewallPolicies | null) => {
-    set(internalFirewallPolicies$, policies);
-  },
-);
-
-const fetchZeroJobFirewallPolicies$ = command(
-  async ({ get, set }, _signal: AbortSignal) => {
-    const name = get(internalAgentName$);
-    if (!name) {
-      return;
-    }
-
-    try {
-      const client = get(zeroClient$)(zeroAgentsByIdContract);
-      const result = await client.get({ params: { id: name } });
-      if (result.status !== 200) {
-        L.warn(`Failed to fetch firewall policies (${result.status})`);
-        return;
-      }
-
-      set(internalFirewallPolicies$, result.body.firewallPolicies ?? null);
-    } catch (error) {
-      throwIfAbort(error);
-      L.error("Failed to fetch firewall policies:", error);
-    }
-  },
-);
-
-// ---------------------------------------------------------------------------
-// Agent schedule
-// ---------------------------------------------------------------------------
-
-interface ScheduleItem {
-  id: string;
-  agentId: string;
-  name: string;
-  enabled: boolean;
-  triggerType: "cron" | "once" | "loop";
-  cronExpression: string | null;
-  atTime: string | null;
-  intervalSeconds: number | null;
-  timezone: string;
-  prompt: string;
-  description: string | null;
-  createdAt: string;
-}
-
-// ---------------------------------------------------------------------------
-// Schedule time string conversion
-// ---------------------------------------------------------------------------
-
-function cronToTimeString(cron: string): string {
-  const parts = cron.split(" ");
-  const minute = Number(parts[0]);
-  const hour = Number(parts[1]);
-  const dayOfMonth = parts[2] ?? "*";
-  const dayOfWeek = parts[4] ?? "*";
-
-  const ampm = hour >= 12 ? "PM" : "AM";
-  const h12 = hour === 0 ? 12 : hour > 12 ? hour - 12 : hour;
-  const timeStr = `${h12}:${String(minute).padStart(2, "0")} ${ampm}`;
-
-  if (dayOfWeek === "1-5") {
-    return `Every weekday at ${timeStr}`;
-  }
-  if (dayOfMonth !== "*") {
-    return `Every month on day ${dayOfMonth} at ${timeStr}`;
-  }
-  if (dayOfWeek !== "*") {
-    const dayNames: Record<string, string> = {
-      "0": "Sunday",
-      "1": "Monday",
-      "2": "Tuesday",
-      "3": "Wednesday",
-      "4": "Thursday",
-      "5": "Friday",
-      "6": "Saturday",
-    };
-    const days = dayOfWeek
-      .split(",")
-      .map((d) => {
-        return dayNames[d];
-      })
-      .filter(Boolean);
-    if (days.length > 0) {
-      return `Every week on ${days.join(", ")} at ${timeStr}`;
-    }
-    return `Every week at ${timeStr}`;
-  }
-  return `Every day at ${timeStr}`;
-}
-
-function scheduleToTimeString(s: ScheduleItem): string {
-  if (s.triggerType === "loop" && s.intervalSeconds !== null) {
-    if (s.intervalSeconds % 60 === 0) {
-      return `Every ${s.intervalSeconds / 60} minutes`;
-    }
-    return `Every ${s.intervalSeconds} seconds`;
-  }
-  if (s.triggerType === "once" && s.atTime) {
-    const at = new Date(s.atTime);
-    const date = `${at.getFullYear()}-${String(at.getMonth() + 1).padStart(2, "0")}-${String(at.getDate()).padStart(2, "0")}`;
-    const hour = at.getHours();
-    const minute = at.getMinutes();
-    const ampm = hour >= 12 ? "PM" : "AM";
-    const h12 = hour === 0 ? 12 : hour > 12 ? hour - 12 : hour;
-    return `Once on ${date} at ${h12}:${String(minute).padStart(2, "0")} ${ampm}`;
-  }
-  if (s.cronExpression) {
-    return cronToTimeString(s.cronExpression);
-  }
-  return "Scheduled";
-}
-
-// ---------------------------------------------------------------------------
-// Schedule state
-// ---------------------------------------------------------------------------
-
-interface ZeroJobScheduleState {
-  schedules: ScheduleItem[];
-  error: string | null;
-}
-
-const scheduleState$ = state<ZeroJobScheduleState>({
-  schedules: [],
-  error: null,
-});
-
-const scheduleLoaded$ = state(false);
-
-export const zeroJobScheduleEntries$ = computed((get) => {
-  const items = get(scheduleState$).schedules;
-  return [...items]
-    .sort((a, b) => {
-      return b.createdAt.localeCompare(a.createdAt);
-    })
-    .map((s): ScheduleEntry => {
-      return {
-        id: s.id,
-        time: scheduleToTimeString(s),
-        prompt: s.prompt,
-        description: s.description,
-        enabled: s.enabled,
-        name: s.name,
-        timezone: s.timezone,
-        intervalSeconds: s.intervalSeconds,
-      };
-    });
-});
-
-export const zeroJobScheduleLoading$ = computed((get) => {
-  return !get(scheduleLoaded$);
-});
-
-export const zeroJobScheduleError$ = computed((get) => {
-  return get(scheduleState$).error;
-});
-
-const fetchZeroJobSchedule$ = command(
-  async ({ get, set }, _signal: AbortSignal) => {
-    const detail = get(zeroJobDetail$);
-    if (!detail) {
-      return;
-    }
-
-    try {
-      const client = get(zeroClient$)(zeroSchedulesMainContract);
-      const result = await client.list();
-      if (result.status !== 200) {
-        throw new Error(`Failed to fetch schedules (${result.status})`);
-      }
-
-      const agentSchedules = result.body.schedules.filter((s) => {
-        return s.agentId === detail.agentId;
-      });
-      set(scheduleState$, { schedules: agentSchedules, error: null });
-      set(scheduleLoaded$, true);
-    } catch (error) {
-      throwIfAbort(error);
-      L.error("Failed to fetch schedules:", error);
-      set(scheduleState$, {
-        schedules: [],
-        error:
-          error instanceof Error ? error.message : "Failed to load schedules",
-      });
-      set(scheduleLoaded$, true);
-    }
-  },
-);
-
-// ---------------------------------------------------------------------------
-// Schedule CRUD
-// ---------------------------------------------------------------------------
-
-export interface ZeroJobScheduleSaveParams {
-  prompt: string;
-  description?: string;
-  freq: string;
-  date: string;
-  hour: number;
-  minute: number;
-  timezone: string;
-  intervalSeconds: number;
-  dayOfWeek?: string;
-  dayOfMonth?: string;
-  editName?: string;
-}
-
-export const saveZeroJobSchedule$ = command(
-  async (
-    { get, set },
-    params: ZeroJobScheduleSaveParams,
-    signal: AbortSignal,
-  ) => {
-    const detail = get(zeroJobDetail$);
-    if (!detail) {
-      throw new Error("No agent detail loaded");
-    }
-
-    const scheduleName = params.editName ?? `zero-${Date.now().toString(36)}`;
-
-    const base = {
-      agentId: detail.agentId,
-      name: scheduleName,
-      timezone: params.timezone,
-      prompt: params.prompt.trim(),
-      ...(params.description && { description: params.description.trim() }),
-      enabled: true,
-    };
-
-    let body: ScheduleBody;
-
-    if (params.freq === "every_n_minutes") {
-      body = { ...base, intervalSeconds: params.intervalSeconds };
-    } else if (params.freq === "once") {
-      if (
-        isAtTimePast(params.date, String(params.hour), String(params.minute))
-      ) {
-        throw new Error("Scheduled time must be in the future");
-      }
-      const atTime = buildAtTime(
-        params.date,
-        String(params.hour),
-        String(params.minute),
-      );
-      body = { ...base, atTime };
-    } else if (params.freq === "now") {
-      body = { ...base, atTime: new Date().toISOString() };
-    } else {
-      const freqMap: Record<string, CronTimeOption> = {
-        every_weekday: "every-weekday",
-        every_day: "every-day",
-        every_week: "every-week",
-        every_month: "every-month",
-      };
-      const timeOption = freqMap[params.freq];
-      if (!timeOption) {
-        throw new Error(`Unknown schedule frequency: ${params.freq}`);
-      }
-      const cronExpression = buildCronExpression({
-        timeOption,
-        hour: String(params.hour),
-        minute: String(params.minute),
-        dayOfWeek: params.dayOfWeek,
-        dayOfMonth: params.dayOfMonth,
-      });
-      body = { ...base, cronExpression };
-    }
-
-    const client = get(zeroClient$)(zeroSchedulesMainContract);
-    const result = await client.deploy({ body });
-    signal.throwIfAborted();
-
-    if (result.status !== 200 && result.status !== 201) {
-      const detail =
-        result.status === 400 ||
-        result.status === 401 ||
-        result.status === 403 ||
-        result.status === 404
-          ? result.body.error.message
-          : `Save failed (${result.status})`;
-      throw new Error(detail);
-    }
-
-    toast.success(params.editName ? "Schedule updated" : "Schedule created");
-    await set(fetchZeroJobSchedule$, signal);
-  },
-);
-
-export const toggleZeroJobScheduleEnabled$ = command(
-  async (
-    { get, set },
-    params: { name: string; enabled: boolean },
-    signal: AbortSignal,
-  ) => {
-    const detail = get(zeroJobDetail$);
-    if (!detail) {
-      throw new Error("No agent detail loaded");
-    }
-
-    const client = get(zeroClient$)(zeroSchedulesEnableContract);
-    const action = params.enabled ? "enable" : "disable";
-    const result = await client[action]({
-      params: { name: params.name },
-      body: { agentId: detail.agentId },
-    });
-    signal.throwIfAborted();
-
-    if (result.status !== 200) {
-      const message = `Failed to ${action} schedule (${result.status})`;
-      toast.error(message);
-      throw new Error(message);
-    }
-
-    // Optimistic update: patch the local schedule state instead of refetching
-    const current = get(scheduleState$);
-    set(scheduleState$, {
-      ...current,
-      schedules: current.schedules.map((s) => {
-        return s.name === params.name ? { ...s, enabled: params.enabled } : s;
-      }),
-    });
-  },
-);
-
-export const deleteZeroJobSchedule$ = command(
-  async ({ get, set }, scheduleName: string, signal: AbortSignal) => {
-    const detail = get(zeroJobDetail$);
-    if (!detail) {
-      throw new Error("No agent detail loaded");
-    }
-
-    const client = get(zeroClient$)(zeroSchedulesByNameContract);
-    const result = await client.delete({
-      params: { name: scheduleName },
-      query: { agentId: detail.agentId },
-    });
-    signal.throwIfAborted();
-
-    if (result.status !== 204) {
-      const msg =
-        result.status === 401 || result.status === 403 || result.status === 404
-          ? result.body.error.message
-          : `status ${result.status}`;
-      throw new Error(`Delete failed: ${msg}`);
-    }
-
-    toast.success("Schedule deleted");
-    await set(fetchZeroJobSchedule$, signal);
-  },
-);
-
-// ---------------------------------------------------------------------------
-// Delete agent
-// ---------------------------------------------------------------------------
-
-export const deleteZeroJobAgent$ = command(
-  async ({ get, set }, signal: AbortSignal) => {
-    const detail = get(zeroJobDetail$);
-    if (!detail) {
-      throw new Error("No agent detail loaded");
-    }
-
-    const client = get(zeroClient$)(zeroAgentsByIdContract);
-    const result = await client.delete({ params: { id: detail.agentId } });
-    signal.throwIfAborted();
-
-    if (result.status !== 204) {
-      const msg =
-        result.status === 401 || result.status === 403 || result.status === 404
-          ? result.body.error.message
-          : `status ${result.status}`;
-      throw new Error(`Delete failed: ${msg}`);
-    }
-
-    toast.success("Agent deleted");
-    set(reloadAgents$);
-  },
-);
-
-// ---------------------------------------------------------------------------
-// Combined fetch — loads detail, then instructions + schedule in parallel
-// ---------------------------------------------------------------------------
-
-export const fetchZeroJobData$ = command(
-  async ({ set }, agentName: string, signal: AbortSignal) => {
-    // Reset all state so the skeleton screen shows while loading new data
-    set(detailState$, { detail: null, loading: false, error: null });
-    set(instructionsState$, {
-      instructions: null,
-      loading: false,
-      error: null,
-    });
-    set(scheduleState$, { schedules: [], error: null });
-    set(scheduleLoaded$, false);
-    set(editedContent$, null);
-    set(internalAddedConnectors$, null);
-    set(userConnectorPermissionsState$, {
-      enabledTypes: [],
-      loading: false,
-      error: null,
-    });
-    set(internalBuildError$, null);
-    set(jobBuilding$, false);
-    set(internalSaving$, false);
-    set(internalActiveTab$, getInitialTab());
-    set(internalFirewallPolicies$, null);
-
-    set(setZeroJobAgentName$, agentName);
-    await set(fetchZeroJobDetail$, signal);
-    await Promise.all([
-      set(fetchZeroJobInstructions$, signal),
-      set(fetchZeroJobSchedule$, signal),
-      set(fetchZeroJobFirewallPolicies$, signal),
-      set(fetchZeroJobUserConnectors$, signal),
-    ]);
-  },
-);
+export * from "./job-detail/index.ts";
